# docs/_generated/openapi/sir-convert-a-lot-v2.openapi.json
{
  "components": {
    "schemas": {
...
      "DigiExamMigrationBundleManifestV2": {
        "additionalProperties": false,
        "description": "Terminal artifact bundle manifest for DigiExam migration jobs.",
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/DigiExamMigrationBundleArtifactEntryV2"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "bundle_status": {
            "enum": [
              "complete",
              "partial",
              "needs_review",
              "failed"
            ],
            "title": "Bundle Status",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "manual_follow_up": {
            "$ref": "#/components/schemas/DigiExamMigrationBundleManualFollowUpV2"
          },
          "readiness": {
            "$ref": "#/components/schemas/DigiExamMigrationBundleReadinessSummaryV2"
          },
          "retention": {
            "$ref": "#/components/schemas/DigiExamMigrationBundleRetentionV2"
          },
          "schema_version": {
            "const": "digiexam_migration_bundle_v3",
            "title": "Schema Version",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/DigiExamMigrationBundleSourceV2"
          },
          "source_binding": {
            "$ref": "#/components/schemas/DigiExamMigrationBundleSourceBindingV2"
          },
          "warnings": {
            "$ref": "#/components/schemas/DigiExamMigrationBundleWarningsV2"
          }
        },
        "required": [
          "schema_version",
          "job_id",
          "source",
          "bundle_status",
          "retention",
          "artifacts",
          "manual_follow_up",
          "readiness",
          "source_binding",
          "warnings"
        ],
        "title": "DigiExamMigrationBundleManifestV2",
        "type": "object"
      },
      "DigiExamMigrationBundleManualFollowUpV2": {
        "additionalProperties": false,
        "description": "Manual follow-up summary in current DigiExam migration bundle manifests.",
        "properties": {
          "artifact_key": {
            "const": "manual_follow_up_report",
            "title": "Artifact Key",
            "type": "string"
          },
          "count": {
            "minimum": 0,
            "title": "Count",
            "type": "integer"
          },
          "required": {
            "title": "Required",
            "type": "boolean"
          }
        },
        "required": [
          "required",
          "artifact_key",
          "count"
        ],
        "title": "DigiExamMigrationBundleManualFollowUpV2",
        "type": "object"
      },
      "DigiExamMigrationBundleReadinessSummaryV2": {
        "additionalProperties": false,
        "description": "Readiness summary pointing consumers to the authoritative report.",
        "properties": {
          "artifact_key": {
            "const": "target_readiness_report",
            "title": "Artifact Key",
            "type": "string"
          },
          "exportable_targets": {
            "items": {
              "type": "string"
            },
            "title": "Exportable Targets",
            "type": "array"
          },
          "review_required": {
            "title": "Review Required",
            "type": "boolean"
          }
        },
        "required": [
          "artifact_key",
          "review_required"
        ],
        "title": "DigiExamMigrationBundleReadinessSummaryV2",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobPendingResultResponseV2": {
        "additionalProperties": false,
        "description": "Response payload when a v2 result request is made before completion.",
        "properties": {
          "api_version": {
            "const": "v2",
            "default": "v2",
            "title": "Api Version",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobPendingResultResponseV2",
        "type": "object"
      },
...
      "JobResultResponseV2": {
        "additionalProperties": false,
        "description": "Response payload for v2 successful result retrieval.",
        "properties": {
          "api_version": {
            "const": "v2",
            "default": "v2",
            "title": "Api Version",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/ResultPayloadV2"
          },
          "status": {
            "const": "succeeded",
            "default": "succeeded",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "result"
        ],
        "title": "JobResultResponseV2",
        "type": "object"
      },
...
      "JobStatus": {
        "description": "Job status lifecycle states.",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "canceled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "PdfCheckpointV2": {
        "additionalProperties": false,
        "description": "Durable checkpoint state for chunked PDF-to-markdown execution.",
        "properties": {
          "chunk_size_pages": {
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Chunk Size Pages",
            "type": "integer"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/PdfChunkRecordV2"
            },
            "title": "Chunks",
            "type": "array"
          },
          "failed_pages": {
            "minimum": 0.0,
            "title": "Failed Pages",
            "type": "integer"
          },
          "job_id": {
            "minLength": 1,
            "title": "Job Id",
            "type": "string"
          },
          "processed_pages": {
            "minimum": 0.0,
            "title": "Processed Pages",
            "type": "integer"
          },
          "schema_version": {
            "const": "v2_pdf_checkpoint_v2",
            "default": "v2_pdf_checkpoint_v2",
            "title": "Schema Version",
            "type": "string"
          },
          "total_pages": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Pages"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "updated_at",
          "chunk_size_pages",
          "processed_pages",
          "failed_pages"
        ],
        "title": "PdfCheckpointV2",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Sir Convert-a-Lot API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/convert/jobs/{job_id}/artifact": {
      "get": {
        "operationId": "get_artifact_v2_convert_jobs__job_id__artifact_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPendingResultResponseV2"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact"
      }
    },
    "/v2/convert/jobs/{job_id}/artifact/partial": {
      "get": {
        "operationId": "get_partial_artifact_v2_convert_jobs__job_id__artifact_partial_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPendingResultResponseV2"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Partial Artifact"
      }
    },
    "/v2/convert/jobs/{job_id}/artifacts": {
      "get": {
        "operationId": "get_artifact_bundle_manifest_v2_convert_jobs__job_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigiExamMigrationBundleManifestV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPendingResultResponseV2"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact Bundle Manifest"
      }
    },
    "/v2/convert/jobs/{job_id}/artifacts/{artifact_key}": {
      "get": {
        "operationId": "get_named_artifact_v2_convert_jobs__job_id__artifacts__artifact_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_key",
            "required": true,
            "schema": {
              "title": "Artifact Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPendingResultResponseV2"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Named Artifact"
      }
    },
...
    "/v2/convert/jobs/{job_id}/checkpoint": {
      "get": {
        "operationId": "get_checkpoint_v2_convert_jobs__job_id__checkpoint_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PdfCheckpointV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPendingResultResponseV2"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Checkpoint"
      }
    },
...
    "/v2/convert/jobs/{job_id}/result": {
      "get": {
        "operationId": "get_result_v2_convert_jobs__job_id__result_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResultResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPendingResultResponseV2"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Result"
      }
    },
...
  }
}